+++ api/src/infrastructure/data/typeorm/entities/highlighted-category.entity.ts
