+++ api/src/routes/products.route.ts
@@ -1,13 +1,13 @@
 import express, {NextFunction, Request, Response} from 'express';
 import ProductUseCase from '../application/usecases/product.usecase';
-import SQLiteCategoryRepository from "../infrastructure/data/sqllite/category.repository.sqlite";
-import SQLiteProductRepository from "../infrastructure/data/sqllite/product.repository.sqlite";
+import TypeORMCategoryRepository from "../infrastructure/data/typeorm/category.repository.typeorm";
+import TypeORMProductRepository from "../infrastructure/data/typeorm/product.repository.typeorm";
 import ProductService from "../core/services/product.service";
 import {AppDataSource} from "../data-source";
 
 const router = express.Router();
-const categoryRepository = new SQLiteCategoryRepository(AppDataSource);
-const productRepository = new SQLiteProductRepository(AppDataSource);
+const categoryRepository = new TypeORMCategoryRepository(AppDataSource);
+const productRepository = new TypeORMProductRepository(AppDataSource);
 const productService = new ProductService(productRepository);
 const productUseCase = new ProductUseCase(productService, productRepository, categoryRepository);
